Rename file_system to filesystem in file names
